+++ spark-mongodb/src/main/scala/com/stratio/datasource/mongodb/reader/MongodbReader.scala
@@ -24,7 +24,7 @@ import com.stratio.datasource.mongodb.partitioner.MongodbPartition
 import org.apache.spark.Partition
 import org.apache.spark.sql.sources._
 import org.apache.spark.sql.types.UTF8String
-import scala.util.Try
+import scala.util.{Failure, Success, Try}
 import java.util.regex.Pattern
 
 /**
@@ -78,10 +78,14 @@ class MongodbReader(
             MongoCredential.createCredential(user,database,password)},
         config.get[MongodbSSLOptions](MongodbConfig.SSLOptions), config.properties.filterKeys(_.contains(MongodbConfig.ListMongoClientOptions))))
 
+
+      val emptyFilter = MongoDBObject(List())
+      val filter = Try(queryPartition(filters)).getOrElse(emptyFilter)
+
       dbCursor = (for {
         client <- mongoClient
         collection <- Option(client(config(MongodbConfig.Database))(config(MongodbConfig.Collection)))
-        dbCursor <- Option(collection.find(queryPartition(filters), selectFields(requiredColumns)))
+        dbCursor <- Option(collection.find(filter, selectFields(requiredColumns)))
       } yield {
           mongoPartition.partitionRange.minKey.foreach(min => dbCursor.addSpecial("$min", min))
           mongoPartition.partitionRange.maxKey.foreach(max => dbCursor.addSpecial("$max", max))
@@ -138,7 +142,6 @@ class MongodbReader(
 
       queryBuilder.get
     }
-
     filtersToDBObject(filters)
   }